Alterações nas atividade2 Java – POO: Classes e Objetos atualizado e corrigido.
