--- src/generation/dadosfuncionario/TestaFuncionario.java
@@ -1,11 +1,0 @@
-package generation.dadosfuncionario;
-
-public class TestaFuncionario {
-
-    public static void main(String[] args) {
-    Funcionario funcionario1 = new Funcionario(54698,"Marcos Lucas Silva",
-            "Desenvolvedor Java","Desenvolver aplicação em Java","Tecnologia da informação");
-
-     funcionario1.visualizar();
-    }
-}
